Ditched copy-pasted code
Should function the same same as the last rev in this branch. Added global variable emergency_tilt which is ordinarily 0, and +- 45 in cases where Safety_Override would otherwise have activated the right or left thruster to avoid a close horizontal obstacle.

Occasional "Elvis has left the building" issue where it goes too high over the mountains and off the top of the map.