struct State, frame_count
Added a struct to store state variables, and a global frame counter. Initialize an array of them on the first frame, from within the Lander_Control function. Not sure if we are expected to free that memory (how would we know when to do it anyway, since "winning" and ending the program is a black box to our code?)